Refactor the merge sort dumper to allow reuse of its disk read/write ablities
